Added support for signed numbers to avoid negative heights

+++ dblock.go
@@ -157,7 +157,7 @@ func GetDBlockHead() (string, error) {
 }
 
 // ReplayDBlockFromHeight requests DBlock states to be emitted over the LiveFeed API
-func ReplayDBlockFromHeight(startheight uint32, endheight uint32) (*JSON2Response, error) {
+func ReplayDBlockFromHeight(startheight int64, endheight int64) (*JSON2Response, error) {
 	params := replayRequest{StartHeight: startheight, EndHeight: endheight}
 	req := NewJSON2Request("replay-from-height", APICounter(), params)
 	resp, err := factomdRequest(req)
